requirements.txt
python-telegram-bot==20.7
pygame==2.5.2
python-dotenv==1.0.0
mutagen==1.47.0
httpx[socks]==0.25.2
httpx[socks]==0.25.2